addresses issues with prior commit for file test_install.sh

+++ tests/test_install.sh
@@ -1,10 +1,14 @@
-#!/bin/sh
+#!/bin/bash
+set -ueo pipefail
 
-##+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+
-##
-##  Usage: ./test_install.sh [install | uninstall | upgrade]
-##
-##+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+
+if [[ "$#" -ne 1 ]]; then
+  echo "##+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+"
+  echo "##"
+  echo "##  Usage: ./test_install.sh [install | uninstall | upgrade]"
+  echo "##"
+  echo "##+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+"
+  exit 1
+fi
 
 clear
 
@@ -12,57 +16,73 @@ clear
 # $CODESEALER_TOKEN=
 
 # Public Helm Repo
-export CODESEALER_HELM_REPO=tfarinacci/codesealer-helm/main/
-# CODESEALER_HELM_REPO=helm.github.codesealer.com
+export CODESEALER_HELM_REPO=https://code-sealer.github.io/helm-charts
+export CODESEALER_HELM_CHART=codesealer/codesealer
 
-# Name of release
+# Version of release
 export RELEASE_VER="1"
 
+# Installation specific  exports
 export INGRESS_NAMESPACE=ingress-nginx
 export INGRESS_DEPLOYMENT=ingress-nginx-${RELEASE_VER}-controller
 export INGRESS_PORT=443
-
 export REDIS_NAMESPACE=redis
 
 if [[ "$1" == "install" ]]; then
 
-  echo "\n########################################################################################"
+  echo "########################################################################################"
   echo "#  Do you wish to install NGINX Ingress Controller to access the OWASP Juice Shop Application?"
   echo "#  "
   echo "#  Documentation: https://github.com/kubernetes/ingress-nginx/tree/main/charts/ingress-nginx"
   echo "########################################################################################"
-  read -p 'Install NGINX Ingress Controller [y/n]: '
-  if [ $REPLY == 'y' ]; then
+  read -r -p 'Install NGINX Ingress Controller [y/n]: '
+  if [[ ${REPLY} == 'y' ]]; then
+    echo "########################################################################################"
+    echo "#  Waiting for NGINX Ingress Controller to start"
+    echo "########################################################################################"      
+
     helm upgrade --install ingress-nginx-${RELEASE_VER} ingress-nginx \
     --repo https://kubernetes.github.io/ingress-nginx \
-    --namespace $INGRESS_NAMESPACE --create-namespace \
+    --namespace ${INGRESS_NAMESPACE} --create-namespace \
     --set controller.hostPort.enabled=true \
     --set controller.service.type=LoadBalancer \
     --set controller.publishService.enabled=false \
     --set controller.extraArgs.publish-status-address=localhost \
     --wait --timeout=60s
   else
-    echo "\n########################################################################################"
+    echo "########################################################################################"
     echo "#  Skipping NGINX Ingress Controller installation"
     echo "########################################################################################"
   fi
 
-  echo "\n########################################################################################"
+  echo "########################################################################################"
   echo "#  Install OWASP Juice Shop Application"
   echo "#  "
   echo "#  Documentation: https://artifacthub.io/packages/helm/securecodebox/juice-shop"
   echo "########################################################################################"
-  read -s -t 10 -p '?Press any key to continue.'
+  read -r -p 'Install OWASP Juice Shop [y/n]: '
 
-  helm repo add securecodebox https://charts.securecodebox.io/
-  helm install juice-shop-${RELEASE_VER} securecodebox/juice-shop --namespace juice-shop --create-namespace \
-    --set ingress.enabled=true \
-    --set "ingress.hosts[0].host=localhost,ingress.hosts[0].paths[0].path=/" \
-    --set "ingress.tls[0].hosts[0]=localhost,ingress.tls[0].secretName=" \
-    --set ingress.pathType=Prefix \
-    --wait --timeout=60s
+  if [[ ${REPLY} == 'y' ]]; then
+    helm repo add securecodebox https://charts.securecodebox.io/
 
-  echo "\n########################################################################################"
+    echo "########################################################################################"
+    echo "#  Waiting for Juice Shop to start"
+    echo "########################################################################################"      
+
+    helm install juice-shop-${RELEASE_VER} securecodebox/juice-shop --namespace juice-shop --create-namespace \
+      --set ingress.enabled=true \
+      --set "ingress.hosts[0].host=localhost,ingress.hosts[0].paths[0].path=/" \
+      --set "ingress.tls[0].hosts[0]=localhost,ingress.tls[0].secretName=" \
+      --set ingress.pathType=Prefix \
+      --wait --timeout=60s
+
+  else
+    echo "########################################################################################"
+    echo "#  Skipping Juice Shop installation"
+    echo "########################################################################################"
+  fi
+
+  echo "########################################################################################"
   echo "#  Do you wish to install Redis?"
   echo "#  "
   echo "#   This is required by Codesealer.  Either use this installation helm chart"
@@ -71,8 +91,12 @@ if [[ "$1" == "install" ]]; then
   echo "#   -- Using github.com/bitnami/charts repository"
   echo "#      Documentation: https://github.com/bitnami/charts/blob/main/bitnami/redis/README.md"
   echo "########################################################################################"
-  read -p 'Install Redis in single master mode [y/n]: '
-  if [ $REPLY == 'y' ]; then
+  read -r -p 'Install Redis in single master mode [y/n]: '
+  if [[ ${REPLY} == 'y' ]]; then
+
+    echo "########################################################################################"
+    echo "#  Waiting for Redis to start"
+    echo "########################################################################################"      
     helm install redis-${RELEASE_VER} oci://registry-1.docker.io/bitnamicharts/redis \
     --namespace ${REDIS_NAMESPACE} --create-namespace \
     --set auth.enabled=true \
@@ -81,27 +105,28 @@ if [[ "$1" == "install" ]]; then
 
     # --set sentinel.enabled=false \
     # --set auth.sentinel=false \
+
   else
-    echo "\n########################################################################################"
-    echo "#  Skipping Redis installation"
+    echo "########################################################################################"
+    echo "# Skipping Redis installation"
     echo "########################################################################################"
   fi
 
-  echo "\n########################################################################################"
-  echo "#  Install Codesealer"
   echo "########################################################################################"
-  read -s -t 10 -p '?Press any key to continue.'
+  echo "# Install Codesealer"
+  echo "########################################################################################"
+  read -r -s -p 'Press any key to continue.'
   # Install Codesealer helm repo
-  helm repo add codesealer https://raw.githubusercontent.com/${CODESEALER_HELM_REPO}
+  helm repo add codesealer ${CODESEALER_HELM_REPO}
 
   # Get the Redis password
   export REDIS_PASSWORD=$(kubectl get secret --namespace ${REDIS_NAMESPACE} redis-${RELEASE_VER} -o jsonpath="{.data.redis-password}" | base64 -d)
-  echo "Redis password: ${REDIS_PASSWORD}"
-  read -s -t 10 -p '?Press any key to continue.'
-
-  # Install Codsealer
-  # We can only write to the redis master service
-  helm install codesealer-${RELEASE_VER} codesealer/codesealer --create-namespace --namespace codesealer-system \
+  echo "########################################################################################"
+  echo "# Redis password: ${REDIS_PASSWORD}"
+  echo "# "
+  echo "# Waiting for Codesealer to start"
+  echo "########################################################################################"      
+  helm install codesealer-${RELEASE_VER} ${CODESEALER_HELM_CHART} --create-namespace --namespace codesealer-system \
     --set codesealerToken=${CODESEALER_TOKEN} \
     --set worker.ingress.namespace=${INGRESS_NAMESPACE} \
     --set worker.ingress.deployment=${INGRESS_DEPLOYMENT} \
@@ -126,69 +151,77 @@ if [[ "$1" == "install" ]]; then
     # --set "ingress.tls[0].hosts[0]=core-manager.local,ingress.tls[0].secretName=ingress-tls" \
     # --set "ingress.hosts[0].host=core-manager.local,ingress.hosts[0].paths[0].path=/,ingress.hosts[0].paths[0].pathType=Prefix"
 
-  echo "\n########################################################################################"
-  echo "#  Activate Codesealer by applying labels and annotations"
   echo "########################################################################################"
-  read -s -p '?Press any key to continue.'
+  echo "#  Activate Codesealer by applying labels and annotations:"
+  echo "# "
+  echo "# $ kubectl label ns ${INGRESS_NAMESPACE} codesealer.com/webhook=enabled"
+  echo "# "
+  echo "# $ kubectl patch deployment ${INGRESS_DEPLOYMENT} -n ${INGRESS_NAMESPACE} -p '{"spec": {"template":{"metadata":{"annotations":{"codesealer.com/injection":"enabled"}}}} }'"
+  echo "# "
+  echo "########################################################################################"
+  read -r -s -p 'Press any key to continue.'
 
-  # Enable Codesealer
-  echo "\n########################################################################################"
-  echo "  $ kubectl label ns $INGRESS_NAMESPACE codesealer.com/webhook=enabled"
-  echo "\n"
-  kubectl label ns $INGRESS_NAMESPACE codesealer.com/webhook=enabled
-
-  echo "\n"
-  echo "  $ kubectl patch deployment $INGRESS_DEPLOYMENT -n $INGRESS_NAMESPACE -p '{"spec": {"template":{"metadata":{"annotations":{"codesealer.com/injection":"enabled"}}}} }'"
-  echo "\n"
-  kubectl patch deployment $INGRESS_DEPLOYMENT -n $INGRESS_NAMESPACE -p '{"spec": {"template":{"metadata":{"annotations":{"codesealer.com/injection":"enabled"}}}} }'
-  echo "\n########################################################################################"
+  kubectl label ns ${INGRESS_NAMESPACE} codesealer.com/webhook=enabled
+  echo "# "
+  kubectl patch deployment ${INGRESS_DEPLOYMENT} -n ${INGRESS_NAMESPACE} -p '{"spec": {"template":{"metadata":{"annotations":{"codesealer.com/injection":"enabled"}}}} }'
   
-  echo "\n########################################################################################"
+  echo "########################################################################################"
   echo "#  Restart NGINX Ingress Controller"
   echo "########################################################################################"
-  read -s -t 5 -p '?Press any key to continue.'
-  kubectl scale deployment $INGRESS_DEPLOYMENT  --namespace $INGRESS_NAMESPACE --replicas=0
+  read -r -s -p 'Press any key to continue.'
+
+  kubectl scale deployment ${INGRESS_DEPLOYMENT}  --namespace ${INGRESS_NAMESPACE} --replicas=0
+  echo "########################################################################################"
+  echo "#  Waiting for NGINX Ingress Controller to restart"
+  echo "########################################################################################"  
   sleep 20   
-  kubectl scale deployment $INGRESS_DEPLOYMENT  --namespace $INGRESS_NAMESPACE --replicas=1
+  kubectl scale deployment ${INGRESS_DEPLOYMENT}  --namespace ${INGRESS_NAMESPACE} --replicas=1
 
 elif [[ "$1" == "uninstall" ]]; then
 
-  echo "\n########################################################################################"
+  echo "########################################################################################"
   echo "#  Uninstall Codesealer"
   echo "########################################################################################"
   helm uninstall codesealer-${RELEASE_VER} --namespace codesealer-system
   helm repo remove codesealer
   kubectl delete namespace codesealer-system
 
-  echo "\n########################################################################################"
-  echo "#  Uninstall OWASP Juice Shop Application"
   echo "########################################################################################"
-  helm uninstall juice-shop-${RELEASE_VER} --namespace juice-shop
-  helm repo remove securecodebox
-  kubectl delete namespace juice-shop
+  echo "#  Do you wish to uninstall OWASP Juice Shop Application?"
+  echo "########################################################################################"
+  read -r -p 'Uninstall OWASP Juice Shop [y/n]: '
+  if [[ ${REPLY} == 'y' ]]; then
+    helm uninstall juice-shop-${RELEASE_VER} --namespace juice-shop
+    helm repo remove securecodebox
+    kubectl delete namespace juice-shop
+  else
+    echo "########################################################################################"
+    echo "#  Skipping Juice Shop uninstall"
+    echo "########################################################################################"
+  fi
 
-  echo "\n########################################################################################"
+  echo "########################################################################################"
   echo "#  Do you wish to uninstall Redis?"
   echo "########################################################################################"
-  read -p 'Uninstall Redis [y/n]: '
+  read -r -p 'Uninstall Redis [y/n]: '
   if [ $REPLY == 'y' ]; then
     helm uninstall redis-${RELEASE_VER} --namespace redis 
     kubectl delete namespace redis
   else
-    echo "\n########################################################################################"
+    echo "########################################################################################"
     echo "#  Skipping Redis uninstall"
     echo "########################################################################################"
   fi
 
-  echo "\n########################################################################################"
+  echo "########################################################################################"
   echo "#  Do you wish to uninstall NGINX Ingress Controller?"
   echo "########################################################################################"
-  read -p 'Uninstall NGINX Ingress Controller [y/n]: '
+  read -r -p 'Uninstall NGINX Ingress Controller [y/n]: '
   if [ $REPLY == 'y' ]; then
-    helm uninstall ingress-nginx-${RELEASE_VER} --namespace $INGRESS_NAMESPACE
-    # kubectl delete namespace $INGRESS_NAMESPACE
+    helm uninstall ingress-nginx-${RELEASE_VER} --namespace ${INGRESS_NAMESPACE}
+    # kubectl delete namespace ${INGRESS_NAMESPACE}
   else
-    echo "\n########################################################################################"
+    echo "########################################################################################"
     echo "#  Skipping NGINX Ingress Controller uninstall"
     echo "########################################################################################"
   fi
@@ -198,11 +231,11 @@ elif [[ "$1" == "upgrade" ]]; then
   # Get the Redis password
   export REDIS_PASSWORD=$(kubectl get secret --namespace ${REDIS_NAMESPACE} redis -o jsonpath="{.data.redis-password}" | base64 -d)
 
-  echo "\n########################################################################################"
+  echo "########################################################################################"
   echo "#  Upgrade Codesealer Release"
   echo "########################################################################################"
   helm repo update codesealer
-  helm upgrade codesealer-${RELEASE_VER} codesealer/codesealer --namespace codesealer-system \
+  helm upgrade codesealer-${RELEASE_VER} ${CODESEALER_HELM_CHART} --namespace codesealer-system \
     --set codesealerToken=${CODESEALER_TOKEN} \
     --set worker.ingress.namespace=${INGRESS_NAMESPACE} \
     --set worker.ingress.deployment=${INGRESS_DEPLOYMENT} \
@@ -218,31 +251,34 @@ elif [[ "$1" == "upgrade" ]]; then
     --set worker.config.endpoint.crs.paranoiaLevel=1 \
     --wait --timeout=90s
 
-  echo "\n########################################################################################"
+  echo "########################################################################################"
   echo "#  Upgrade Codesealer"
   echo "########################################################################################"
-  read -s -t 10 -p '?Press any key to continue.'
+  read -r -s -p 'Press any key to continue.'
   kubectl rollout restart deployments --namespace codesealer-system
 
-  echo "\n########################################################################################"
+  echo "########################################################################################"
   echo "#  Please wait 1 minute..."
   echo "########################################################################################"
   sleep 60
 
-  echo "\n########################################################################################"
+  echo "########################################################################################"
   echo "#  Upgrade Codesealer Worker Sidecar for NGINX Ingress Controller"
   echo "########################################################################################"
-  read -s -t 30 -p '?Press any key to continue.'
-  # kubectl rollout restart deployments --namespace $INGRESS_NAMESPACE
-  kubectl scale deployment $INGRESS_DEPLOYMENT  --namespace $INGRESS_NAMESPACE --replicas=0
+  read -r -s -p 'Press any key to continue.'
+  # kubectl rollout restart deployments --namespace ${INGRESS_NAMESPACE}
+  kubectl scale deployment ${INGRESS_DEPLOYMENT}  --namespace ${INGRESS_NAMESPACE} --replicas=0
   sleep 20   
-  kubectl scale deployment $INGRESS_DEPLOYMENT  --namespace $INGRESS_NAMESPACE --replicas=1
+  kubectl scale deployment ${INGRESS_DEPLOYMENT}  --namespace ${INGRESS_NAMESPACE} --replicas=1
 
 else
 
-  echo "\n########################################################################################"
-  echo "#  Invalid arguement: must install, upgrade, or uninstall"
-  echo "########################################################################################"
-
+  echo "##+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+"
+  echo "##"
+  echo "##  Invalid arguement!"
+  echo "##"
+  echo "##  Usage: ./test_install.sh [install | uninstall | upgrade]"
+  echo "##"
+  echo "##+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+"
+  exit 1
 fi
-
